# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnlyPump API",
    "version": "1.0.0",
    "description": "Versioned REST gateway for the OnlyPump web and mobile applications. Read-only token intelligence today; see ARCHITECTURE.md for what's planned next."
  },
  "servers": [
    {
      "url": "https://api.onlypump.me/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Supabase access token (or, for internal/admin callers, an API_KEYS token)."
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RobinhoodTokenListResponse": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscoveredToken"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "type": "integer"
          },
          "observedAt": {
            "type": "string"
          }
        },
        "required": [
          "tokens",
          "nextCursor",
          "total",
          "observedAt"
        ]
      },
...
      "RobinhoodStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "LIVE",
              "LAGGING",
              "DEGRADED",
              "REORG_RECOVERY",
              "UNAVAILABLE"
            ]
          },
          "discovery": {
            "$ref": "#/components/schemas/SourceHealthDetail"
          },
          "trades": {
            "$ref": "#/components/schemas/SourceHealthDetail"
          },
          "observedAt": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "discovery",
          "trades",
          "observedAt"
        ]
      },
...
      "RobinhoodTokenDetailResponse": {
        "type": "object",
        "properties": {
          "token": {
            "$ref": "#/components/schemas/DiscoveredToken"
          },
          "trades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChainTrade"
            }
          },
          "observedAt": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "trades",
          "observedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/tokens/robinhood": {
      "get": {
        "summary": "Recently discovered Pons/Robinhood Chain tokens, newest first (Phase 7B.4). Phase 7D.4: server-side lifecycle (all, bonding, graduated, almost-bonded, trending) and search filters with a filtered total, sorting by newest, market cap, liquidity, bonding progress or 1h market-cap change, a verified quote asset and the live on-chain market snapshot per row.",
        "tags": [
          "robinhood-chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "bonding",
                "graduated",
                "almost-bonded",
                "trending"
              ],
              "default": "all"
            },
            "required": false,
            "name": "lifecycle",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "marketCap",
                "liquidity",
                "progress",
                "change1h"
              ]
            },
            "required": false,
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "required": false,
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Discovered token list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobinhoodTokenListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/robinhood/status": {
      "get": {
        "summary": "Backend-owned Pons/Robinhood Chain ingestion source-health projection (Phase 7B.5A) — LIVE/LAGGING/DEGRADED/REORG_RECOVERY/UNAVAILABLE, never inferred client-side.",
        "tags": [
          "robinhood-chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion source health",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobinhoodStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/robinhood/{tokenAddress}": {
      "get": {
        "summary": "One discovered Robinhood Chain token's detail with its recent trades (Phase 7B.4).",
        "tags": [
          "robinhood-chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$",
              "example": "0x055650555Be80649397084Cd3f8a09b4350e8612"
            },
            "required": true,
            "name": "tokenAddress",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Token detail with trades",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobinhoodTokenDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/robinhood/{tokenAddress}/candles": {
      "get": {
        "summary": "Materialized OHLCV candle history for a Robinhood/Pons token (Phase 7B.5B) — reads PostgreSQL only, never Robinhood RPC inline.",
        "tags": [
          "robinhood-chain",
          "candles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$",
              "example": "0x055650555Be80649397084Cd3f8a09b4350e8612"
            },
            "required": true,
            "name": "tokenAddress",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Candle history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tokens/robinhood/{tokenAddress}/pool": {
      "get": {
        "deprecated": true,
        "summary": "Deprecated by /market-evidence (Phase 7D.3.2), which is block-pinned, covers bonding curves and ERC-20 pair assets, and separates raw liquidity from depth. Live Uniswap V4 pool evidence for a graduated Pons V2 token (Phase 7D.3 §5). Native-denominated only — no USD conversion exists for this chain. Always 200: a token with no V4 pool returns status UNAVAILABLE with a reason code.",
        "tags": [
          "robinhood-chain"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$",
              "example": "0x055650555Be80649397084Cd3f8a09b4350e8612"
            },
            "required": true,
            "name": "tokenAddress",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Pool evidence, or an explicit unavailable reason",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolEvidenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}